Handle platform-specific player files
