+++ src/screens/EventDetailsScreen/index.tsx
@@ -71,35 +71,30 @@ const EventDetailsScreen: React.FC<TEventDetalsScreenProps> = ({ route }) => {
   );
 
   if (isBMPlayerShowing) {
-    return (
-    <View style={styles.root}>
-      <RohText style={styles.rootText} bold>
-        iOS bitmovin player coming soon
-      </RohText>
-    </View> 
-    // <Player
-    //   autoPlay
-    //   configuration={{
-    //     url: 'https://video-ingestor-output-bucket.s3.eu-west-1.amazonaws.com/6565/manifest.m3u8',
-    //     poster:
-    //       'https://actualites.music-opera.com/wp-content/uploads/2019/09/14OPENING-superJumbo.jpg',
-    //   }}
-    //   title="event title"
-    //   subtitle="some event subtitle"
-    //   onClose={closePlayer}
-    //   analytics={{
-    //     videoId: 'blahblahblah',
-    //     title: 'Some video title',
-    //     experiment: 'ROH TV app',
-    //     customData1: '',
-    //     customData2: '',
-    //     customData3: '',
-    //     customData4: '',
-    //     customData5: '',
-    //     customData6: '',
-    //     customData7: '',
-    //   }}
-    // />
+    return (  
+    <Player
+      autoPlay
+      configuration={{
+        url: 'https://video-ingestor-output-bucket.s3.eu-west-1.amazonaws.com/6565/manifest.m3u8',
+        poster:
+          'https://actualites.music-opera.com/wp-content/uploads/2019/09/14OPENING-superJumbo.jpg',
+      }}
+      title="event title"
+      subtitle="some event subtitle"
+      onClose={closePlayer}
+      analytics={{
+        videoId: 'blahblahblah',
+        title: 'Some video title',
+        experiment: 'ROH TV app',
+        customData1: '',
+        customData2: '',
+        customData3: '',
+        customData4: '',
+        customData5: '',
+        customData6: '',
+        customData7: '',
+      }}
+    />
     );
   }
   return (
@@ -121,10 +116,6 @@ const styles = StyleSheet.create({
     flex: 1,
   },
   root: { flex: 1, alignItems: 'center', justifyContent: 'center' },
-  rootText: {
-    color: 'white',
-    fontSize: scaleSize(48),
-  },
 });
 
 export default EventDetailsScreen;
